+++ allennlp/data/dataloader.py
@@ -1,5 +1,4 @@
 from typing import List, Dict, Union, Iterator
-import warnings
 
 import torch
 from torch.utils import data
@@ -7,7 +6,6 @@ from torch.utils import data
 from allennlp.common.registrable import Registrable
 from allennlp.common.lazy import Lazy
 from allennlp.data.instance import Instance
-from allennlp.data.dataset_readers.dataset_reader import AllennlpLazyDataset
 from allennlp.data.batch import Batch
 from allennlp.data.samplers import Sampler, BatchSampler
 
@@ -87,13 +85,6 @@ class PyTorchDataLoader(data.DataLoader, DataLoader):
         multiprocessing_context: str = None,
         batches_per_epoch: int = None,
     ):
-        if num_workers and isinstance(dataset, AllennlpLazyDataset):
-            warnings.warn(
-                "Using multi-process data loading with a lazy dataset could lead to "
-                "deadlocks with certain tokenizers. See:\n"
-                "  https://github.com/allenai/allennlp/issues/4330\n",
-                UserWarning,
-            )
         super().__init__(
             dataset=dataset,
             batch_size=batch_size,