+++ tests/training/trainer_test.py
@@ -990,6 +990,43 @@ class TestTrainer(TrainerTestBase):
         expected_calls = [epoch for epoch in range(-1, 4)]
         assert trainer.epoch_callback_calls == expected_calls
 
+    def test_total_loss_is_average_of_batch_loss(self):
+
+        batches_per_epoch = 3
+
+        data_loader_custom_epoch_lazy = PyTorchDataLoader(
+            self.instances_lazy,
+            batch_size=2,
+            collate_fn=allennlp_collate,
+            batches_per_epoch=batches_per_epoch,
+        )
+
+        class FakeBatchCallback(BatchCallback):
+            def __call__(
+                self,
+                trainer: "GradientDescentTrainer",
+                batch_inputs: List[List[TensorDict]],
+                batch_outputs: List[Dict[str, Any]],
+                epoch: int,
+                batch_number: int,
+                is_training: bool,
+                is_master: bool,
+            ) -> None:
+                if not hasattr(trainer, "batch_losses"):
+                    trainer.batch_losses = []  # type: ignore
+                trainer.batch_losses.append(batch_outputs[0]["loss"].item())  # type: ignore
+
+        trainer = GradientDescentTrainer(
+            self.model,
+            self.optimizer,
+            data_loader_custom_epoch_lazy,
+            num_epochs=1,
+            batch_callbacks=[FakeBatchCallback()],
+        )
+        metrics = trainer.train()
+
+        assert metrics["training_loss"] == float(sum(trainer.batch_losses) / batches_per_epoch)
+
 
 class TestApexTrainer(TrainerTestBase):
     @requires_gpu