+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed unnecessary warning about deadlocks in `DataLoader`.
 - Use slower tqdm intervals when output is being piped or redirected.
-- Fixed testing models that only return a loss when they are in training mode
+- Fixed testing models that only return a loss when they are in training mode.
+- Fixed a bug in `FromParams` that causes silent failure in case of the parameter type being Optional[Union[...]].
 
 ### Added